requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
chardet==4.0.0
colorama==0.4.4
commonmark==0.9.1
discord.py==1.6.0
idna==3.1
multidict==5.1.0
Pygments==2.7.3
rich==9.6.1
typing-extensions==3.7.4.3
yarl==1.6.3
